Split OGR helpers and polyline structs
