+++ CMakeLists.txt
@@ -1,5 +1,7 @@
 cmake_minimum_required(VERSION 3.8)
 
+project(visvalingam)
+
 # Enable C++11
 set(CMAKE_CXX_STANDARD 11)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -8,7 +10,8 @@ set(CMAKE_POSITION_INDEPENDENT_CODE ON)
 find_package(GDAL REQUIRED)
 
 add_library(visvalingam_algorithm
-  src/geo_types.cpp
+  src/ogr.cpp
+  src/polyline.cpp
   src/visvalingam_algorithm.cpp
 )
 target_link_libraries(visvalingam_algorithm PUBLIC ${GDAL_LIBRARY})
